Cleanup, version 2.0
