+++ CHANGELOG.md
@@ -1,173 +1,7 @@
-## 1.4 2025-01-09
-
-- Compiler: add support defer
-
-## 1.3 2024-12-16
-
-- Update dependencies
-
-## 1.2 2023-08-21
-
-- Update dependencies
-- Implement template fragments
-
-## 1.1 2023-07-03
-
-- Add direct iteration using the `_for` attribute
-
-## 1.0 2023-03-30
-
-- Add support for Array attribute values
-
-## 0.29 2023-03-11
-
-- Add Tilt integration (#15)
-
-## 0.28 2023-03-11
-
-- Add `HTML#import_map`, `HTML#js_module` methods
-
-## 0.27 2023-01-19
-
-- Fix rendering of HTML void element tags
-
-## 0.26 2023-01-13
-
-- Add support for namespaced local extensions using `#extend`
-
-## 0.25 2023-01-12
-
-- Implement `#def_tag` for defining custom tags inline
-
-## 0.24 2022-03-19
-
-- Fix usage of const components (#13)
-- Fix formatting of HTML/XML attributes for non-string values
-
-## 0.23 2022-02-15
-
-- Remove unused `Encoding` module
-- Add SOAP extension (#11, thanks [@aemadrid](https://github.com/aemadrid))
-
-## 0.22 2022-02-14
-
-- Fix behaviour of call to `#p` in an extension (#10)
-
-## 0.21 2022-02-13
-
-- Refactor and improve documentation
-
-## 0.20 2022-02-13
-
-- Add support for XML namespaced tags and attributes (#9)
-- Move and refactor HTML/XML common code to Tags module
-
-## 0.19 2022-02-05
-
-- Rename `P2::Component` to `P2::Template`
-
-## 0.18 2022-02-04
-
-- Cleanup and update examples
-- Fix behaviour of #emit with block
-- Improve README
-
-## 0.17 2022-01-23
-
-- Refactor markdown code, add `P2.markdown` method (#8)
-
-## 0.16 2022-01-23
-
-- Implement JSON templating (#7)
-- Add support for MIME types (#6)
-- Change entrypoint from `Kernel#H`, `Kernel#X` to `P2.html`, `.xml` (#5)
-
-## 0.15 2022-01-20
-
-- Fix tag method line reference
-- Don't clobber ArgumentError exception
-
-## 0.14 2022-01-19
-
-- Add support for #emit_yield in applied component (#4)
-
-## 0.13 2022-01-19
-
-- Add support for partial parameter application (#3)
-
-## 0.12 2022-01-06
-
-- Improve documentation
-- Add `Renderer#tag` method
-- Add `HTML#style`, `HTML#script` methods
-
-## 0.11 2022-01-04
-
-- Add deferred evaluation
-
-## 0.10.1 2021-12-25
-
-- Fix tag rendering with empty text in Ruby 3.0
-
-## 0.10 2021-12-25
-
-- Add support for extensions
-
-## 0.9 2021-12-23
-
-- Add support for emitting Markdown
-- Add support for passing proc as argument to `#H` and `#X`
-- Deprecate `Encoding` module
-
-## 0.8.1 2021-12-22
-
-- Fix gemspec
-
-## 0.8 2021-12-22
-
-- Cleanup and refactor code
-- Add P2.xml global method for XML templates
-- Make `Component` a descendant of `Proc`
-- Introduce new component API
-- Rename Rubyoshka to P2
-- Convert underscores to dashes for tag  and attribute names (@jaredcwhite)
-
-## 0.7 2021-09-29
-
-- Add `#emit_yield` for rendering layouts
-- Add experimental template compilation (WIP)
-
-## 0.6.1 2021-03-03
-
-- Remove support for Ruby 2.6
-
-## 0.6 2021-03-03
-
-- Fix Rubyoshka on Ruby 3.0
-- Refactor and add more tests
-
-## 0.5 2021-02-27
-
-- Add support for rendering XML
-- Add Rubyoshka.component method
-- Remove Modulation dependency
-
-## 0.4 2019-02-05
-
-- Add support for emitting component modules
-
-## 0.3 2019-01-13
-
-- Implement caching
-- Improve performance
-- Handle attributes with `false` value correctly
-
-## 0.2 2019-01-07
-
-- Better documentation
-- Fix #text
-- Add local context
-
-## 0.1 2019-01-06
-
-- First working version
+## 2.0 2025-08-07
+
+- Passes all HTML, compilation tests from Papercraft
+- Automatic compilation
+- Plain procs/lambdas as templates
+- Remove everything not having to do with HTML
+- P2: compiled functional templates - they're super fast!
