+++ lib/p2/compiler.rb
@@ -544,57 +544,4 @@ module P2
       emit("; __buffer__ = __orig_buffer__; __parts__.each { it.is_a?(Proc) ? it.() : (__buffer__ << it) }")
     end
   end
-
-  module AuxMethods
-    def format_tag(tag)
-      tag.to_s.gsub('_', '-')
-    end
-
-    def format_html_attr_key(tag)
-      tag.to_s.tr('_', '-')
-    end
-  
-    def format_html_attrs(attrs)
-      attrs.each_with_object(+'') do |(k, v), html|
-        case v
-        when nil, false
-        when true
-          html << ' ' if !html.empty?
-          html << format_html_attr_key(k)
-        else
-          html << ' ' if !html.empty?
-          v = v.join(' ') if v.is_a?(Array)
-          html << "#{format_html_attr_key(k)}=\"#{v}\""
-        end
-      end
-    end
-
-    def render_emit_call(o, *a, **b, &block)
-      case o
-      when nil
-        # do nothing
-      when ::Proc
-        o.render(*a, **b, &block)
-      else
-        o.to_s
-      end
-    end
-
-    def translate_backtrace(e, source_map)
-      re = /^(#{source_map[:compiled_fn]}\:(\d+))/
-      source_fn = source_map[:source_fn]
-      backtrace = e.backtrace.map {
-        if (m = it.match(re))
-          line = m[2].to_i
-          source_line = source_map[line] || "?(#{line})"
-          it.sub(m[1], "#{source_fn}:#{source_line}")
-        else
-          it
-        end
-      }
-      e.set_backtrace(backtrace)
-    end
-  end
-  
-  P2.extend(AuxMethods)
 end
