+++ lib/p2/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module P2
-  VERSION = '1.4'
+  VERSION = '2.0'
 end
