+++ test/fixtures/compiler/emit_template_source.rb
@@ -1,9 +1,9 @@
-pr1 = P2.html { p 'foo' }
-pr2 = P2.html { |x| p x }
+pr1 = -> { p 'foo' }
+pr2 = ->(x) { p x }
 
 ->() {
   emit pr1
   emit pr2, 42
   br
-  emit P2.html { q 'bar' }
+  emit -> { q 'bar' }
 }
